+++ ggEngine/ggEngine/Shape.h
@@ -5,7 +5,7 @@
 namespace ggEngine {
 	class Shape: public GGObject {
 	public:
-		virtual float GetArea() = 0;
+		virtual double GetArea() = 0;
 		virtual void Transform(Matrix transformMatrix) = 0;
 	};
 }
